backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
joblib==1.5.3
numpy==2.3.5
pandas==2.3.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.4
threadpoolctl==3.6.0
tzdata==2025.3
